feat(header): Mejoras en el navbar v1
