+++ src/app/app.routes.ts
@@ -1,5 +1,6 @@
 import { Routes } from '@angular/router';
 import {HomeComponent} from './pages/home/home.component';
+import {CommentComponent} from './component/comment/comment.component';
 
 export const routes: Routes = [
   { path: '', redirectTo: '/home', pathMatch: 'full' },
@@ -7,5 +8,6 @@ export const routes: Routes = [
   // Define otras rutas aquí
   { path: 'login', component: HomeComponent },
   { path: 'register', component: HomeComponent },
-  { path: '**', redirectTo: '/home' }, // Ruta para manejar errores 404
+  { path: 'comment', component: CommentComponent },
+  //{ path: '**', redirectTo: '/home' }, // Ruta para manejar errores 404
 ];
