+++ src/app/component/carousel/carousel.component.scss
@@ -1,0 +1,3 @@
+.h-96 {
+  height: 35rem/* 384px */;
+}
