--- src/app/shared/carousel/carousel.component.html
@@ -1,71 +1,0 @@
-<div class="relative w-full h-96 overflow-hidden shadow-lg">
-  <div
-    class="flex transition-transform duration-700 ease-in-out"
-    [style.transform]="'translateX(-' + currentSlide * 100 + '%)'"
-  >
-    <div *ngFor="let slide of slides; let i = index" class="min-w-full relative">
-      <img [src]="slide.imageUrl" [alt]="slide.title" class="w-full h-96 object-cover">
-      <div class="absolute inset-0 bg-black bg-opacity-40 flex items-center justify-center">
-        <div class="text-center text-white p-4">
-          <h2 class="text-3xl font-bold mb-4">{{ slide.title }}</h2>
-          <p class="mb-6">{{ slide.description }}</p>
-          <button
-            (click)="goToSlide(i)"
-            class="bg-blue-600 hover:bg-blue-700 text-white font-bold py-2 px-6 rounded-full transition duration-300"
-          >
-            {{ slide.ctaText }}
-          </button>
-        </div>
-      </div>
-    </div>
-  </div>
-
-  <button
-    (click)="prevSlide()"
-    aria-label="Previous slide"
-    class="absolute left-4 top-1/2 transform -translate-y-1/2 bg-white bg-opacity-50 hover:bg-opacity-75 rounded-full p-2 focus:outline-none focus:ring-2 focus:ring-blue-500"
-  >
-    <svg class="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-      <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M15 19l-7-7 7-7"></path>
-    </svg>
-  </button>
-
-  <button
-    (click)="nextSlide()"
-    aria-label="Next slide"
-    class="absolute right-4 top-1/2 transform -translate-y-1/2 bg-white bg-opacity-50 hover:bg-opacity-75 rounded-full p-2 focus:outline-none focus:ring-2 focus:ring-blue-500"
-  >
-    <svg class="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-      <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9 5l7 7-7 7"></path>
-    </svg>
-  </button>
-
-  <div class="absolute bottom-4 left-1/2 transform -translate-x-1/2 flex space-x-2">
-    <button
-      *ngFor="let slide of slides; let i = index"
-      (click)="goToSlide(i)"
-      [attr.aria-label]="'Go to slide ' + (i + 1)"
-      class="w-3 h-3 rounded-full focus:outline-none focus:ring-2 focus:ring-blue-500"
-      [ngClass]="{ 'bg-white': currentSlide === i, 'bg-white/50': currentSlide !== i }"
-    ></button>
-  </div>
-
-  <button
-    (click)="toggleAutoplay()"
-    class="absolute top-4 right-4 bg-white bg-opacity-50 hover:bg-opacity-75 p-2 rounded-full focus:outline-none focus:ring-2 focus:ring-blue-500"
-  >
-    <svg *ngIf="!isPlaying" class="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-      <path
-        stroke-linecap="round"
-        stroke-linejoin="round"
-        stroke-width="2"
-        d="M14.752 11.168l-3.197-2.132A1 1 0 0010 9.87v4.263a1 1 0 001.555.832l3.197-2.132a1 1 0 000-1.664z"
-      />
-    </svg>
-    <svg *ngIf="isPlaying" class="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-      <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M10 9v6m4-6v6"></path>
-    </svg>
-  </button>
-</div>
-
-
