--- src/app/shared/carousel/carousel.component.scss
@@ -1,3 +1,0 @@
-.h-96 {
-  height: 35rem/* 384px */;
-}
